refactor(scan): extract shared receipt-date helpers to src/lib/receipt-date.ts
Removes duplication of parseLocalDate and checkReceiptDate between the scan
and breakdown routes. No behavior change.
